refactor: remove subnet= config directive, auto-build from rules
Address PR #61 review feedback:

- Remove config-level subnet= directive from parseConfig()
- Auto-build SubnetContext from rule-level subnet: match keys
- Remove baseline router/netmask block from Handler4() (rules handle this)
- Add AddSubnetCIDROnly() to SubnetContext for CIDR-only registration
- Move SubnetPoolManager tests to separate pool_test.go
- Update tests for auto-built SubnetContext behavior
- Update coredhcp.yaml and rules.md documentation

CoreSMD now becomes relay-aware natively when rules contain subnet:
match keys, eliminating the redundancy between subnet= and rules.
